+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 v0.16.0
 
 - Return `chromsizes` as a single array in beddb `tileset_info`
+- [BREAKING] Remove the `chrom_names` and `chrom_sizes` fields in the beddb tileset info
 
 v0.15.4